requirements.txt

# Requirements for the project
fastapi>=0.111
uvicorn[standard]>=0.30
httpx>=0.27
pydantic>=2.7
python-dotenv>=1.0
ruff>=0.5
mypy>=1.11
types-requests
pre-commit>=3.8

# For LLMs and vector databases
langchain>=0.3
langgraph>=0.2
langchain-google-genai


# For geospatial data handling
pandas
openai
pyproj


# For testing
pytest>=8.3
pytest-cov>=5.0
